requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
openai
python-multipart
stripe
python-telegram-bot
gTTS
pydub
SpeechRecognition
vocab
requests